requirements.txt
progress==1.6
requests==2.32.3
matplotlib==3.10.3
python-socketio[client,asyncio_client]==5.13.0
pyyaml==6.0.2
tabulate==0.9.0
pandas==2.3.0
numpy==2.3.1
eventlet==0.40.1